+++ compose/src/commonMain/kotlin/ComposeNode.kt
@@ -1,8 +1,6 @@
 package dev.inmo.navigation.compose
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import dev.inmo.kslog.common.logger
 import dev.inmo.micro_utils.coroutines.SpecialMutableStateFlow
 import dev.inmo.navigation.core.NavigationChain
 import dev.inmo.navigation.core.NavigationNode
@@ -10,7 +8,6 @@ import dev.inmo.navigation.core.NavigationNodeId
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlin.jvm.JvmName
 
 /**
  * Provides [onDraw] open function which will be called by the navigation system to draw content in the place it added
@@ -45,7 +42,7 @@ abstract class ComposeNode<Config : Base, Base>(
      */
     @Composable
     protected open fun SubchainsHost(filter: (NavigationChain<Base>) -> Boolean) {
-        defaultSubchainsHandling(filter)
+        SubchainsHandling(filter)
     }
 
     @Composable