+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Navigation Changelog
 
+## 0.5.6
+
+* `Versions`:
+  * `MicroUtils`: `0.22.3` -> `0.22.4`
+* `Core`:
+  * Add default config for `Empty` node (just empty config object)
+  * Add default factory for `Empty` node, based on default config
+  * Add `NavigationNodeFactory.Typed` default node factory with inline factory fun
+* `Compose`:
+  * Add opportunity to setup initial chains and nodes in `initNavigation`
+  * A lot of fixes in Compose Navigation Extensions
+
 ## 0.5.5
 
 * `Compose`: